Complete project of RAG based document chatbot

+++ src/entities/base/_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Column, DateTime, func
+from sqlalchemy import Integer, Column, DateTime, func
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -8,7 +8,7 @@ class BaseModel_(Base):
     __abstract__ = True  # This model should not be instantiated directly
     __table_args__ = {"extend_existing": True}  # Allow table extension
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True, nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
     created_at = Column(DateTime, server_default=func.now(), nullable=False)
     updated_at = Column(
         DateTime, server_default=func.now(), onupdate=func.now(), nullable=False
